Rename metafunctions module to core.
In an attempt to make package structure cleaner
