+++ metafunctions/core.py
