+++ metafunctions/tests/test_function_merge.py
@@ -1,7 +1,7 @@
 import operator
 
-from metafunctions.metafunctions import FunctionMerge
-from metafunctions.metafunctions import SimpleFunction
+from metafunctions.core import FunctionMerge
+from metafunctions.core import SimpleFunction
 from metafunctions.tests.util import BaseTestCase
